refactor: upper case err
