+++ traflix-web/src/pages/SearchPage/DesktopDefault.module.scss
