+++ traflix-web/src/pages/SearchPage/MobileDefault.module.scss
