+++ traflix-web/src/pages/SearchPage/index.tsx
